change modal messages to comunicate throught service

+++ src/app/heroes/hero-list/hero-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, ViewChild, Inject } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
+import 'rxjs/add/operator/map';
 import { Subscription } from 'rxjs/Subscription';
 
 // models
@@ -13,8 +14,6 @@ import { HeroService } from '../shared/hero.service';
 import { SpinnerService } from '../../core/spinner/spinner.service';
 import { LoggerService } from '../../core/services/logger.service';
 import { MessageService } from '../../modal-message/message.service';
-
-import 'rxjs/add/operator/map';
 
 @Component({
   selector: 'app-hero-list',
@@ -56,23 +55,30 @@ export class HeroListComponent implements OnInit {
 
   delete(id: number) {
     this.heroIdSelected = id;
+    localStorage['action'] = 'DELETE_HERO';
+    localStorage['from'] = 'HERO_LIST';
     this.messageService.showMessage(new Message('Are you sure do you want to delete this Hero?', 'warning', MessageType.CONFIRM));
   }
 
   onOkDelete(value) {
-    if (value) {
-      this.service.delete(this.heroIdSelected).subscribe(res => {
-        if (res.ok) {
-          const index = this.data.findIndex(hero => hero.id === this.heroIdSelected);
-          this.data.splice(index, 1);
-          this.heroIdSelected = null;
-          this.messageService.showMessage(new Message('The super hero was deleted successfully!!', 'success'));
-        } else {
-          this.messageService.showMessage(new Message('Impossible to delete the super hero!'));
-        }
-      }, err => this.messageService.showMessage(new Message('Impossible to delete the super hero!')));
-    } else {
-      this.heroIdSelected = null;
+    if (localStorage['action'] === 'DELETE_HERO' && localStorage['from'] === 'HERO_LIST') {
+      localStorage.removeItem('action');
+      localStorage.removeItem('from');
+
+      if (value) {
+        this.service.delete(this.heroIdSelected).subscribe(res => {
+          if (res.ok) {
+            const index = this.data.findIndex(hero => hero.id === this.heroIdSelected);
+            this.data.splice(index, 1);
+            this.heroIdSelected = null;
+            this.messageService.showMessage(new Message('The super hero was deleted successfully!!', 'success'));
+          } else {
+            this.messageService.showMessage(new Message('Impossible to delete the super hero!'));
+          }
+        }, err => this.messageService.showMessage(new Message('Impossible to delete the super hero!')));
+      } else {
+        this.heroIdSelected = null;
+      }
     }
   }
 }
